Avoid PrefetchOp destructor assertion when not necessary
Summary:
Countless hours were spent debugging why ImageInputOp failed with a cryptic exception P56967302. Turns out, that assertion happened in PrefetchOp destructor, that was triggered when a assertion failed in ImageInputOp constructor. Because of this, the underlying problem was shadowed. I fixed this by not asserting on finalize_ if there is no prefetch thread running, and now the error is clean:

[enforce fail at image_input_op.h:105] scale_ > 0. -1 vs 0. Must provide the scaling factor.

Reviewed By: Yangqing

Differential Revision: D4435105

fbshipit-source-id: 52f85a9fd30eea396c9faca54b6d946fa847b7ff