Sparse momentum update for seq2seq embeddings
Summary: Added SparseMomentumSGDUpdate to NMT training pipeline. Also surfaced and fixed out-of-bounds error in operator stemming from the implicit assumption that gradient slice input would be 2D. Now it is compatible with any dimensions, with indices indexing into the first dimension of param. Added internal checks to ensure that indices are valid.

Differential Revision: D4799697

fbshipit-source-id: 91ea23a6e743cc5337b46fae2821e773067d911e