Create signal failure blobs in constructor, avoid race condition
Summary: andrewwdye caught a sigsegv that happened at Gloo failure signaling function. Turns out workspace->CreateBlob() is not thread safe, and since we are running multiple threads it is likely that many gloo ops fail at once and thus we get a race. Caffe2 ops should actually be created in constructor, so that's what this diff does.

Reviewed By: andrewwdye

Differential Revision: D5139269

fbshipit-source-id: 7eaab3084e4e39543632c628c5e0710225e73b65