Use simple function pointers for memory allocation and deallocation.
Summary:
During the team meeting today Dima and Alex mentioned that the current lambda
function causes slowdown in performance when a large number of alloc and
dealloc happen. My observation is that most of the Delete are actually direct
Delete() function pointers, so I gave it a shot to see if we can reduce
the overhead.

RawAllocDealloc is much fast already, and we observe another 5ns reduction
(12.5%). For TensorAllocDealloc of 32x32 tensors, we are observing 57ns saving
(26%). This is measured on Xeon(R) CPU E5-2660.

Also cleaned up the function interfaces of ShareExternalPointer so we have 2
functions only.

Reviewed By: salexspb, dzhulgakov

Differential Revision: D5801013

fbshipit-source-id: 7068207a43400fa3902bbb3689b3c729e839456c