support model init
Summary:
a parameter can be initialized multiple times in init_net if parameter sharing is enabled. With the original implementation, only the first parameter init will be replaced by pre-trained parameters and the next are still unchanged. This overwrites the initialization with pre-trained parameters.
This diff fixes this issue and also support model init for ads-intent project

Reviewed By: dragonxlwang

Differential Revision: D5991291

fbshipit-source-id: 36173f6239c56bd0d604a77bd94e36072f32faa7