Fix LoadModel() in resnet50_trainer
Summary:
resnet50 trainer will save the 'optimizer_iteration' blob in checkpoints, but loads it i in GPU context. This fails because AtomicIter/Iter expect the blob to be in CPU context. So manually reset the optimizer_iteration in CPU context.

I am thinking of making the iter-operators automatically do this switch, but in the mean time this unbreaks the trainer.

Reviewed By: sf-wind

Differential Revision: D6232626

fbshipit-source-id: da7c183a87803e008f94c86b6574b879c3b76438