fix bug for conv3d Op cpu
There is a bug in ConvOp. SetDeviceTensor function only copies data to tensor when the sizes of the two are different. In the 3d convolution case for video models, img_shape_device_ (NCTWH) is modified only in the first processed example, and for the following examples, it won't get updated, because img_shape_device_.size() == img_shape.size(). However, it should get updated for each example, because T is changing for different videos. It is the same with col_buffer_shape_device_.

In this diff, if any dimension of img_shape_device_ is different from img_shape, img_shape_device_ get updated.